使用策略类实现权限控制
1. 生成策略类php artisan make:policy PostPolicy
2.定义策略，如update,delete
3.注册策略，authprovider中修改
4.策略判断，后端是        $this->authorize('delete',$post);，前端    @can('delete',$post)

+++ laravel54/app/Http/Controllers/PostController.php
@@ -82,6 +82,9 @@ class PostController extends Controller
             'content' => 'required|min:10',
         ]);
 
+        //对用户授权：修改
+        $this->authorize('update',$post);
+
         //更新数据
         $post->title = request('title');
         $post->content = request('content');
@@ -96,8 +99,8 @@ class PostController extends Controller
      */
     public function delete(Post $post)
     {
-        //TODO:用户的权限验证
-
+        //对用户授权：修改
+        $this->authorize('delete',$post);
         $post->delete();
         return redirect('/posts');
     }
