+++ laravel54/resources/views/post/show.blade.php
@@ -5,9 +5,12 @@
     <h2>{{$post->title}}</h2>
 
 
-    <a href="/posts/{{$post->id}}/edit">edit</a>
-    <a href="/posts/{{$post->id}}/delete">delete</a>
-
+    @can('update',$post)
+        <a href="/posts/{{$post->id}}/edit">edit</a>
+    @endcan
+    @can('delete',$post)
+        <a href="/posts/{{$post->id}}/delete">delete</a>
+    @endcan
     <div>
         {!! $post->content !!}
     </div>
